Fixed a compile problem in t/local/ptr_cast_test.c for some gcc 	versions. Reported by "Ryan McGuigan via RT".
git-svn-id: r247

+++ t/local/ptr_cast_test.c
@@ -24,7 +24,7 @@ int main(argc, argv)
   volatile TOTYPE i; /* prevent optimization */
 
   printf("# %s: '%s' len: %ul, '%s' len: %ul.\n", argv[0], FROMTYPESTR,
-	 sizeof(TOTYPE), TOTYPESTR, sizeof(char *));
+	 (int)sizeof(TOTYPE), TOTYPESTR, (int)sizeof(char *));
 
   i = (TOTYPE)bufptr;
   if( ((FROMTYPE)i) != bufptr ) {
